Update explanatory diagram and project vs directory identifiers
We use p for Gradle projects and d for File directories on disk. We
could have used any identifier but choosing the same letter is a bit of
a mind bender for newbies that are not familiar.

We also change the diagram to better match how the code sets up the
projects.